Increased read buffer size
We don't have correct read buffer handling. If a packet length exceeds
our read buffer, we'll only get a partial read and parsing will fail.
The correct logic would be to read more off the connection appending the
data.  Implementing this correctly is quite complex.  Added a FIXME, but
in the meantime we'll increase the size of the buffer to minimize risk.