fix issue #88
updated trap.go so that the oid for the timeticks field is sysUpTime, as
described by Net-SNMP which is what we are emulating:

http://www.net-snmp.org/tutorial/tutorial-5/commands/snmptrap.html

After that change, `go test -test.run` passes!

also changed the TestSendTrap test case to use its own GoSNMP struct
rather than the default so that it plays nice with other test cases.